Added Context API to pass props to RootLayout
